streamlit/requirements.txt
streamlit==1.31.1
httpx==0.25.1
python-dotenv==1.0.0 